+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [1.29.6](https://github.com/brunofgmag/gsx-integrator-client/compare/v1.29.5...v1.29.6) (2026-09-02)
+
+
+* do not repeat a pending completion pick on a snapshot of the same menu ([#124](https://github.com/brunofgmag/gsx-integrator-client/issues/124)) ([ad0d549](https://github.com/brunofgmag/gsx-integrator-client/commit/ad0d549eb748639d843c303aa26150baff2661f4))
+* name what serves the 1L and put the airstair back out when something else stows it ([#123](https://github.com/brunofgmag/gsx-integrator-client/issues/123)) ([c86279b](https://github.com/brunofgmag/gsx-integrator-client/commit/c86279be61ee1c6c7f94982b52862c61d3eb72f3))
+
+
+### Bug Fixes
+
+* **avro-rj:** name what serves the 1L and put the airstair back out when something else stows it ([cad7247](https://github.com/brunofgmag/gsx-integrator-client/commit/cad7247ea42356f9dbb9f8ac4bcbabfe97343227)) ([c86279b](https://github.com/brunofgmag/gsx-integrator-client/commit/c86279be61ee1c6c7f94982b52862c61d3eb72f3))
+* **gsx:** do not repeat a pending completion pick on a snapshot of the same menu ([1ee7e49](https://github.com/brunofgmag/gsx-integrator-client/commit/1ee7e491bd68985a1df9e98b3a8cb4a3a9aa1dcc)) ([ad0d549](https://github.com/brunofgmag/gsx-integrator-client/commit/ad0d549eb748639d843c303aa26150baff2661f4))
+
+## [1.29.5](https://github.com/brunofgmag/gsx-integrator-client/compare/v1.29.4...v1.29.5) (2026-09-02)
+
+
+* hold a door command while the door is still moving, with the Fenix exits measured ([#120](https://github.com/brunofgmag/gsx-integrator-client/issues/120)) ([1b94b00](https://github.com/brunofgmag/gsx-integrator-client/commit/1b94b00383b89241cd6d5b5aeff406e3d4b3b5cb))
+
+
+### Bug Fixes
+
+* **fenix:** watch the three measured exits so door commands wait for the door ([ae227b1](https://github.com/brunofgmag/gsx-integrator-client/commit/ae227b1bf1b565faf83a42f2dbc4501eb3546375)) ([1b94b00](https://github.com/brunofgmag/gsx-integrator-client/commit/1b94b00383b89241cd6d5b5aeff406e3d4b3b5cb))
+* **gsx:** hold a door command while the watched exit is still moving ([391c24b](https://github.com/brunofgmag/gsx-integrator-client/commit/391c24bd59f1bc1f500de03b5d757cfad66e4ed3)) ([1b94b00](https://github.com/brunofgmag/gsx-integrator-client/commit/1b94b00383b89241cd6d5b5aeff406e3d4b3b5cb))
+
 ## [1.29.4](https://github.com/brunofgmag/gsx-integrator-client/compare/v1.29.3...v1.29.4) (2026-09-02)